+++ internal/handler/cmd_query.go
@@ -34,7 +34,7 @@ import (
 //
 // The passed context is canceled when the client connection is closed.
 func (h *Handler) CmdQuery(connCtx context.Context, query *wire.OpQuery) (*wire.OpReply, error) {
-	q, err := bson.TypesDocument(query.Query())
+	q, err := bson.ToDocument(query.Query())
 	if err != nil {
 		return nil, lazyerrors.Error(err)
 	}
@@ -43,7 +43,7 @@ func (h *Handler) CmdQuery(connCtx context.Context, query *wire.OpQuery) (*wire.
 	collection := query.FullCollectionName
 
 	if !strings.HasSuffix(collection, ".$cmd") {
-		return wire.NewOpReply(must.NotFail(bson.ConvertDocument(must.NotFail(types.NewDocument(
+		return wire.NewOpReply(must.NotFail(bson.FromDocument(must.NotFail(types.NewDocument(
 			"$err", "OP_QUERY is no longer supported. The client driver may require an upgrade.",
 			"code", int32(handlererrors.ErrOpQueryCollectionSuffixMissing),
 			"ok", float64(0),
@@ -60,7 +60,7 @@ func (h *Handler) CmdQuery(connCtx context.Context, query *wire.OpQuery) (*wire.
 
 		v, _ := q.Get("speculativeAuthenticate")
 		if v == nil {
-			return wire.NewOpReply(must.NotFail(bson.ConvertDocument(reply)))
+			return wire.NewOpReply(must.NotFail(bson.FromDocument(reply)))
 		}
 
 		authDoc, ok := v.(*types.Document)
@@ -76,7 +76,7 @@ func (h *Handler) CmdQuery(connCtx context.Context, query *wire.OpQuery) (*wire.
 		if err != nil {
 			h.L.DebugContext(connCtx, "No `db` in `speculativeAuthenticate`", logging.Error(err))
 
-			return wire.NewOpReply(must.NotFail(bson.ConvertDocument(reply)))
+			return wire.NewOpReply(must.NotFail(bson.FromDocument(reply)))
 		}
 
 		speculativeAuthenticate, err := h.saslStart(connCtx, dbName, authDoc)
@@ -85,7 +85,7 @@ func (h *Handler) CmdQuery(connCtx context.Context, query *wire.OpQuery) (*wire.
 
 			// unsuccessful speculative authentication leave `speculativeAuthenticate` field unset
 			// and let `saslStart` return an error
-			return wire.NewOpReply(must.NotFail(bson.ConvertDocument(reply)))
+			return wire.NewOpReply(must.NotFail(bson.FromDocument(reply)))
 		}
 
 		h.L.DebugContext(connCtx, "Speculative authentication passed")
@@ -95,7 +95,7 @@ func (h *Handler) CmdQuery(connCtx context.Context, query *wire.OpQuery) (*wire.
 		// saslSupportedMechs is used by the client as default mechanisms if `mechanisms` is unset
 		reply.Set("saslSupportedMechs", must.NotFail(types.NewArray("SCRAM-SHA-1", "SCRAM-SHA-256", "PLAIN")))
 
-		return wire.NewOpReply(must.NotFail(bson.ConvertDocument(reply)))
+		return wire.NewOpReply(must.NotFail(bson.FromDocument(reply)))
 	}
 
 	return nil, handlererrors.NewCommandErrorMsgWithArgument(
